Add context = 0 to gloox::PubSub::ResultHandler
This was also causing a compile time error and is probably due to a
change in the interface of the upnp project. This project now compiles,
but there is some inhertiance in qxmpp that is throwing a warning. It
deals with catching events and I'm not sure how to fix it.